bmp280: build on linux and qurt
